+++ src/ui/display.tsx
@@ -185,7 +185,7 @@ const DisplayApple2 = () => {
   const machineName = handleGetMachineName()
   const slotConfig = handleGetSlotConfig()
   const isApple2Plus = machineName === "APPLE2P"
-  const hasAuxCard = !isApple2Plus && slotConfig[3] === "aux"
+  const hasAuxCard = !isApple2Plus && (slotConfig[3] === "aux" || slotConfig[3] === "vidhd")
   const mem = isApple2Plus ? 64 : (hasAuxCard ? (handleGetMemSize() + 64) : 64)
   const memSize = (mem > 1100) ? ((mem / 1024).toFixed() + " MB") : (mem + " KB")
   const status = <div className="default-font footer-item" translate="no">