+++ src/worker/softswitches.ts
@@ -113,6 +113,7 @@ export const SWITCHES = {
   BSRREADRAM: NewSwitch(0, 0, 0xC012),  // status location, not a switch
   VBLINV: NewSwitch(0, 0, 0xC019),  // vertical blanking status (inverse)
   CASSOUT: NewSwitch(0xC020, 0, 0),  // random value filled in checkSoftSwitches
+  NEWVIDEO: NewSwitch(0, 0, 0xC029), // VidHD / Apple IIGS Super Hi-Res control
   SPEAKER: NewSwitch(0xC030, 0, 0, false, (addr, cycleCount) => {
     memSetC000(0xC030, rand())
     passClickSpeaker(cycleCount)